+++ py/prospect/utilities.py
@@ -52,14 +52,14 @@ vi_flags = [
     # Definition of VI flags
     # Replaces former list viflags = ["Yes","No","Maybe","LowSNR","Bad"]
     # shortlabels for "issue" flags must be a unique single-letter identifier
-    {"label" : "4", "type" : "class", "description" : "Confident classification, two or more secure features"},
-    {"label" : "3", "type" : "class", "description" : "Probable classification, at least one secure feature + continuum; or many weak features"},
-    {"label" : "2", "type" : "class", "description" : "Possible classification, one strong emission feature, but not sure what it is"},
-    {"label" : "1", "type" : "class", "description" : "Unlikely classification, one or some unidentified features"},
-    {"label" : "0", "type" : "class", "description" : "Nothing there"},
-    {"label" : "Bad redshift fit", "shortlabel" : "R", "type" : "issue", "description" : "Misestimation of redshift by pipeline fitter"},
-    {"label" : "Bad spectype fit", "shortlabel" : "C", "type" : "issue", "description" : "Misidentification of spectral category by pipeline fitter, eg. star vs QSO..."},
-    {"label" : "Bad spectrum", "shortlabel" : "S", "type" : "issue", "description" : "Bad spectrum, eg. cosmic / skyline subtraction residuals..."}
+    {"label" : "4", "type" : "class", "description" : "Confident classification: two or more secure features."},
+    {"label" : "3", "type" : "class", "description" : "Probable classification: at least one secure spectral feature + continuum or many weak spectral features."},
+    {"label" : "2", "type" : "class", "description" : "Possible classification: one strong spectral feature but unsure what it is."},
+    {"label" : "1", "type" : "class", "description" : "Unlikely classification: clear signal but features are unidentified."},
+    {"label" : "0", "type" : "class", "description" : "Nothing there, no signal."},
+    {"label" : "Bad redshift fit", "shortlabel" : "R", "type" : "issue", "description" : "Mis-estimation of redshift by the pipeline fitter"},
+    {"label" : "Bad spectype fit", "shortlabel" : "C", "type" : "issue", "description" : "Mis-identification of spectral type from the best-fit pipeline solution; e.g., star vs QSO..."},
+    {"label" : "Bad spectrum", "shortlabel" : "S", "type" : "issue", "description" : "Bad spectrum; e.g. strong cosmic/skyline subtraction residuals."}
 ]
 
 vi_file_fields = [